serverside/planteo/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
defusedxml==0.8.0rc2
Django==5.0.4
django-crontab==0.7.1
django-redis==5.4.0
django-restframework==0.0.1
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
firebase==4.0.1
firebase-admin==6.5.0
fonttools==4.53.0
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.0
msgpack==1.0.8
numpy==2.0.0
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pillow==10.3.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pylibmc==1.6.3
pymemcache==4.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
redis==5.0.7
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.5.0
threadpoolctl==3.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.7.0
